+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Fixed
 
 ## Change
+* gui: Do not overwrite existing applications when creating new ones
 
 # 1.10.0
 ## Added